restructure + use of config file
